Change "AxisType.Dendrogram" into "AxisType.UserDefined" and move the code for DendrogramScale into the Skyline project. Expose the setter for the property Axis.Scale so that the caller can pass in a Scale that they have constructed themselves.
