+++ pwiz_tools/Shared/zedgraph/ZedGraph/Scale.cs
@@ -816,8 +816,6 @@ namespace ZedGraph
 					return new DateAsOrdinalScale( oldScale, _ownerAxis );
 				case AxisType.LinearAsOrdinal:
 					return new LinearAsOrdinalScale( oldScale, _ownerAxis );
-                case AxisType.Dendrogram:
-                    return new DendrogramScale(oldScale, _ownerAxis);
                 default:
 					throw new Exception( "Implementation Error: Invalid AxisType" );
 			}
