+++ pwiz_tools/Skyline/Controls/Clustering/ClusterGraphResults.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using pwiz.Common.Collections;
+using pwiz.Common.DataAnalysis.Clustering;
 using ZedGraph;
 
 namespace pwiz.Skyline.Controls.Clustering
